requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
praw==7.1.0
prawcore==1.5.0
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
